+++ frontend/tiggle/src/components/atoms/index.tsx
@@ -1,0 +1,29 @@
+export { default as CTAButton } from "./CTAButton/CTAButton";
+
+export { default as DatePicker } from "./DatePicker/DatePicker";
+
+export { default as HashTag } from "./HashTag/HashTag";
+
+export { default as Input } from "./Input/Input";
+
+export { default as Loading } from "./Loading/Loading";
+
+export { Menu, MenuItem } from "./Menu/Menu";
+
+export { default as MultiSelect } from "./MultiSelect/MultiSelect";
+
+export { default as ReactionButton } from "./ReactionButton/ReactionButton";
+
+export { default as ReplyToggleButton } from "./ReplyToggleButton/ReplyToggleButton";
+
+export { default as Select } from "./Select/Select";
+
+export { default as SocialLoginButton } from "./SocialLoginButton/SocialLoginButton";
+
+export { default as TextArea } from "./TextArea/TextArea";
+
+export { default as TextButton } from "./TextButton/TextButton";
+
+export { default as TypeTag } from "./typeTag/TypeTag";
+
+export { default as Upload } from "./Upload/Upload";
